# 🚀 Rework unesc for a 63+% performance boost to all of postcss.