@samccone 2021-04-09
Thanks @mathiasbynens and @alexander-akait for the review! All updated and sped up another 10% or so!