@alexander-akait 2021-04-10
@samccone @mathiasbynens Can we do small benchmarks, interesting how it faster/slow