@samccone 2021-04-10
Hi @alexander-akait yes, at the bottom of the original commit message I added a perf section outlining the old and new implementation. 

As far as reusing chrome's implementation with a regex this is possible but we would need to update the license of this project to include chrome's licence as well.  

Finally we can totally do a performance test between chrome's and this implementation. Upon first pass chrome's implementation handles a nice edge case that we do not (invalid unicode), but does not handle a few other decoding cases that the existing test cases require. (Changing this while more correct would be a backwards incompatible change to this core part of postcss which would be a larger undertaking)

I will do some further investigation today! 🚀