@samccone 2021-04-11
Hi @alexander-akait and @mathiasbynens a few updates

1. I updated the implementation to now *new* handle the spec edgecase for 0 hex value escaping, long surrogates, and out of bound value https://drafts.csswg.org/css-syntax/#maximum-allowed-code-point - this is the new commit.
2. I ran new perf tests against chrome's implementation and found that our implementation in this CL is significantly faster. 

The numbers are here:

```
original 
        avg: 4.765ms
        std: 2.5159ms
        max: 35ms
        min: 3ms
        
new 
        avg: 1.76ms
        std: 0.7915ms
        max: 12ms
        min: 1ms
        
chrome 
        avg: 4.383ms
        std: 1.7602ms
        max: 34ms
        min: 3ms
```
        
And the testing methodology can be viewed here https://github.com/samccone/postcss-selector-parser/blob/sjs/perf/perf/index.js#L160-L185