@schuay 2021-04-13
To add some history, slow String.p.replace with global regexp and a callable replacer has been known for a long time, and unfortunately I haven't gotten around to fixing it yet - apologies for that. The previous situation was in fact in a very similar context, see https://github.com/nodejs/node/pull/16986#discussion_r150728733 and crbug.com/v8/7081.

I'm hoping to spend some quality time with regexp later this year. For priorization, it would help if you file a V8 bug report for this, just to point out that this is a real issue, still impacting real cases.

In the meantime, it makes sense to use perf workarounds. I know Leszek has been looking at one based on `RegExp.p.exec` (exec, unlike replace, is actually very fast in the common case), which I suppose he will post here soon.