@LeszekSwirski 2021-04-13
Right, the following:
```
function regexpexec(str) {
    let new_str = "";
    let prev_index = 0;
    let match;
    while ((match = unescapeRegExp.exec(str)) != null) {
        let escaped = match[1];
        let escapedWhitespace = match[2];
        if (match.index > prev_index) {
            new_str += str.slice(prev_index, match.index);
        }

        const high = "0x" + escaped - 0x10000;

        // NaN means non-codepoint
        // Workaround erroneous numeric interpretation of +"0x"
        // eslint-disable-next-line no-self-compare
        new_str += high !== high || escapedWhitespace
            ? escaped
            : high < 0
            ? // BMP codepoint
              String.fromCharCode(high + 0x10000)
            : // Supplemental Plane codepoint (surrogate pair)
              String.fromCharCode(
                  (high >> 10) | 0xd800,
                  (high & 0x3ff) | 0xdc00
              );
        prev_index = unescapeRegExp.lastIndex;
    }
    if (str.length > prev_index) {
        new_str += str.slice(prev_index, str.length);
    }
    return new_str;
}
```

gives me a 50% speedup over this manual string walk:

```
original
        avg: 2.592ms
        std: 0.5528ms
        max: 8ms
        min: 2ms

new
        avg: 1.578ms
        std: 0.5531ms
        max: 7ms
        min: 1ms

chrome
        avg: 2.872ms
        std: 0.4956ms
        max: 8ms
        min: 2ms

Regexp.exec
        avg: 1.002ms
        std: 0.1095ms
        max: 2ms
        min: 0ms
```