@samccone 2021-04-16
@LeszekSwirski unfortunately the solution you listed does not pass all of the test cases added in this change. This is not reflective of your solution being incorrect but rather is more reflective of the existing implementation having multiple bugs that were uncovered and fix during this code-review process.

At this point I see several options

## 1. ignore the fixes to spec bugs that this change fixes and take the solution you wrote 

(pros) 
```
+ much faster than what is in place
```
(cons) 

```
- carries forward the incorrect behavior of unesc forward
```

## 2. Take the current implementation 

(pros) 
```
+ much faster than what is in place 
+ more spec compliant and correct
```

(cons)
```
- not as fast as option 1
```

## 3. (maybe 2.a) follow option 2 and follow up with further perf improvements

(pros)

```
+ helpful to the ecosystem
+ much faster than what is currently released
+ more spec compliant
```

(cons)

```
- we know there is still room for improvement through further refactoring
```

---

@alexander-akait my personal recommendation would be to folllow option 3 with continued community investment to continue to iterate on the optimal solution. 

Please all let me know how you would like to move forward.