@LeszekSwirski 2021-04-21
@samccone Makes sense, although fixing that V8 performance issue of course won't help with the correctness issues you found here.